moved entity manager to component
