+++ homeassistant/components/geonetnz_quakes/const.py
@@ -10,7 +10,12 @@ CONF_MMI = "mmi"
 
 FEED = "feed"
 
+DEFAULT_FILTER_TIME_INTERVAL = timedelta(days=7)
 DEFAULT_MINIMUM_MAGNITUDE = 0.0
 DEFAULT_MMI = 3
 DEFAULT_RADIUS = 50.0
 DEFAULT_SCAN_INTERVAL = timedelta(minutes=5)
+
+SIGNAL_DELETE_ENTITY = "geonetnz_quakes_delete_{}"
+SIGNAL_UPDATE_ENTITY = "geonetnz_quakes_update_{}"
+SIGNAL_STATUS = "geonetnz_quakes_status_{}"
